# src/api/openapi.yaml
openapi: 3.0.0
info:
  title: K5 Platform API
  version: 1.0.0
  description: |
    Comprehensive API for K5 educational content platform with PDF processing,
    bilingual content retrieval, assessment generation, and analytics.
  contact:
    name: K5 Platform Team
    email: support@k5platform.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.k5platform.com/v1
    description: Production server
  - url: https://staging-api.k5platform.com/v1
    description: Staging server
  - url: http://localhost:3000/api/v1
    description: Local development server
...
security:
  - bearerAuth: []
paths:
...
  /pdf/upload:
    post:
      tags:
        - PDF Processing
      summary: Upload and queue PDF
      description: Upload a PDF file for processing with grade level and metadata
      operationId: uploadPdf
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
                - gradeLevel
              properties:
                file:
                  type: string
                  format: binary
                  description: PDF file to upload
                gradeLevel:
                  type: string
                  enum: [K, '1', '2', '3', '4', '5']
                  description: Target grade level
                subject:
                  type: string
                  enum: [math, science, reading, writing, social-studies]
                  description: Subject area
                language:
                  type: string
                  enum: [en, ko, bilingual]
                  default: bilingual
                  description: Content language
                metadata:
                  type: object
                  properties:
                    title:
                      type: string
                    author:
                      type: string
                    publisher:
                      type: string
                    isbn:
                      type: string
      responses:
        '201':
          description: PDF uploaded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                        format: uuid
                        example: 550e8400-e29b-41d4-a716-446655440000
                      filename:
                        type: string
                        example: math-workbook-grade3.pdf
                      status:
                        type: string
                        enum: [queued, processing, completed, failed]
                        example: queued
                      queuePosition:
                        type: integer
                        example: 3
                      estimatedProcessingTime:
                        type: integer
                        description: Estimated time in seconds
                        example: 120
                      uploadedAt:
                        type: string
                        format: date-time
                      gradeLevel:
                        type: string
                        example: '3'
                  message:
                    type: string
                    example: PDF uploaded and queued for processing
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '413':
          description: File too large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                success: false
                error:
                  code: FILE_TOO_LARGE
                  message: File size exceeds maximum limit of 50MB
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
...
  /admin/pdf/{id}:
    delete:
      tags:
        - Admin
      summary: Remove PDF and data
      description: Permanently delete PDF and all associated data
      operationId: deletePdf
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: cascade
          in: query
          schema:
            type: boolean
            default: true
          description: Also delete related content and questions
      responses:
        '200':
          description: PDF deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                    example: PDF and all associated data deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from authentication endpoint
  schemas:
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
  responses:
    BadRequest:
      description: Bad request - Invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error:
              code: VALIDATION_ERROR
              message: Invalid request parameters
              details:
                field: gradeLevel
                issue: Must be one of K, 1, 2, 3, 4, 5
    Unauthorized:
      description: Unauthorized - Invalid or missing authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error:
              code: UNAUTHORIZED
              message: Valid authentication token required
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error:
              code: NOT_FOUND
              message: Requested resource not found
    RateLimitExceeded:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error:
              code: RATE_LIMIT_EXCEEDED
              message: Too many requests. Please try again later.
              details:
                limit: 100
                window: 60
                retryAfter: 45